Moving IconDownloader to gui/
