+++ src/gui/IconDownloader.cpp
