+++ src/gui/IconDownloaderDialog.cpp
@@ -23,9 +23,9 @@
 #include "core/Entry.h"
 #include "core/Global.h"
 #include "core/Group.h"
-#include "core/IconDownloader.h"
 #include "core/Metadata.h"
 #include "core/Tools.h"
+#include "gui/IconDownloader.h"
 #include "gui/IconModels.h"
 #ifdef Q_OS_MACOS
 #include "gui/osutils/macutils/MacUtils.h"
